placeholder

+++ src/ExSql/Syntax/Literal.hs
@@ -6,10 +6,11 @@ module ExSql.Syntax.Literal
     ) where
 
 import Data.Extensible (Member)
+import Data.Int (Int64)
 import ExSql.Syntax.Class
 
 data Literal (g :: * -> *) a where
-    LitInt :: Int -> Literal g Int
+    LitInt :: Int64 -> Literal g Int64
     LitBool :: Bool -> Literal g Bool
 
 instance Hoist Literal where
@@ -18,7 +19,7 @@ instance Hoist Literal where
 
 type Constructor g m a = (Ast g, Monad m, Member (NodeTypes g) Literal) => a -> g m a
 
-int :: Constructor g m Int
+int :: Constructor g m Int64
 int = mkAst . return . LitInt
 
 bool :: Constructor g m Bool
